pyproject.toml
[project]
name = "pocrag"
version = "0.1.0"
description = "POC RAG pour la recommandation d'événements culturels via OpenAgenda"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "altair<5",
    "faiss-cpu>=1.8.0",
    "fastapi>=0.115.0",
    "ipykernel>=6.29.0",
    "ipywidgets>=8.1.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=1.2.1",
    "langchain-mistralai>=0.2.0",
    "matplotlib>=3.8.0",
    "mistralai>=2.0.1",
    "mlflow>=2.17.0",
    "numpy>=1.26.0,<2.0.0",
    "pandas>=2.2.0,<3.0.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.9.0",
    "pytest>=8.3.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.36",
    "streamlit>=1.39.0",
    "uvicorn>=0.30.0",
]